Rename GraphQL folders for clarity
