+++ src/graphql/enums/gameDataType.ts
