+++ src/graphql/enums/genshinElement.ts
