+++ src/graphql/enums/stat.ts
