+++ src/graphql/enums/talentType.ts
