+++ src/graphql/enums/weaponType.ts
