+++ src/graphql/mutations/resolvers/dailyRecord.ts
