+++ src/graphql/resolversMap.ts
@@ -15,17 +15,17 @@ import genderEnumsResolvers from "./enums/gender";
 import heightEnumsResolvers from "./enums/height";
 import regionEnumsResolvers from "./enums/region";
 import statEnumsResolvers from "./enums/stat";
-import characterDataResolvers from "./queries/field/characterData";
+import characterDataResolvers from "./queries/resolvers/field/characterData";
 import queryResolvers from "./queries/query";
-import constellationDataResolvers from "./queries/field/constellationData";
+import constellationDataResolvers from "./queries/resolvers/field/constellationData";
 import genshinElementEnumsResolvers from "./enums/genshinElement";
 import weaponTypeEnumsResolvers from "./enums/weaponType";
 import foodTypeEnumsResolvers from "./enums/foodType";
 import talentTypeEnumsResolvers from "./enums/talentType";
-import foodDataResolvers from "./queries/field/foodData";
-import talentDataResolvers from "./queries/field/talentData";
-import weaponDataResolvers from "./queries/field/weaponData";
-import dailyRecordDataResolvers from "./queries/field/dailyRecordData";
+import foodDataResolvers from "./queries/resolvers/field/foodData";
+import talentDataResolvers from "./queries/resolvers/field/talentData";
+import weaponDataResolvers from "./queries/resolvers/field/weaponData";
+import dailyRecordDataResolvers from "./queries/resolvers/field/dailyRecordData";
 import mutations from "./mutations/mutation";
 import gameDataTypeEnumsResolvers from "./enums/gameDataType";
 // import { getGameData } from "../controllers/teyvatdleGameDataController";
